Create password-gen.py
This code uses the tkinter module to display a prompt asking the user to input the desired length of the password. The simpledialog.askinteger function is used to create the prompt, which has a title of "Password Generator" and a minimum value of 1. The user's input is stored in the length variable.

The program then calls the generate_password function, passing in the length variable as an argument. The function generates a random password using the specified length and returns it.

Finally, the program displays the generated password in a message box using the messagebox.showinfo function, which has a title of "Password Generator" and a message that includes the generated password.